moved slack things into slack package
