+++ slack/message.go
@@ -1,6 +1,12 @@
 package slack
 
-import "encoding/json"
+import (
+	"encoding/json"
+	"fmt"
+	"time"
+
+	"github.com/pkg/errors"
+)
 
 type Message struct {
 	Text            string        `json:"text,omitempty" yaml:"text,omitempty"`
@@ -35,4 +41,103 @@ type HistoryMessage struct {
 func (m HistoryMessage) String() string {
 	ju, _ := json.MarshalIndent(m, "", " ")
 	return string(ju)
+}
+
+type listMessagesResponse struct {
+	OK       bool              `json:"ok"`
+	Error    string            `json:"error"`
+	Messages []*HistoryMessage `json:"messages"`
+	HasMore  bool              `json:"has_more"`
+	PinCount int               `json:"pin_count"`
+}
+
+func (a *API) ListMessages(channelID string) ([]*HistoryMessage, error) {
+	data, err := newRequest(ConversationHistory).
+		addParam("token", a.botToken).
+		addParam("channel", channelID).
+		Get()
+	if err != nil {
+		return nil, errors.Wrap(err, "api requests")
+	}
+	resp := &listMessagesResponse{}
+	if err := json.Unmarshal(data, resp); err != nil {
+		return nil, errors.Wrap(err, "unmarshal")
+	}
+	if !resp.OK {
+		return nil, errors.New(resp.Error)
+	}
+	return resp.Messages, nil
+}
+
+// ListTodayMessages lists messages from today
+func (a *API) ListTodayMessages(channelID string) ([]*HistoryMessage, error) {
+	now := time.Now()
+	today := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, time.UTC)
+	data, err := newRequest(ConversationHistory).
+		addParam("token", a.botToken).
+		addParam("channel", channelID).
+		addParam("oldest", fmt.Sprint(today.Unix())).
+		Get()
+	if err != nil {
+		return nil, errors.Wrap(err, "api requests")
+	}
+	resp := &listMessagesResponse{}
+	if err := json.Unmarshal(data, resp); err != nil {
+		return nil, errors.Wrap(err, "unmarshal")
+	}
+	if !resp.OK {
+		return nil, errors.New(resp.Error)
+	}
+	return resp.Messages, nil
+}
+
+type listDirectMessageChannelsReponse struct {
+	OK       bool       `json:"ok"`
+	Error    string     `json:"error"`
+	Channels []*Channel `json:"ims"`
+}
+
+// ListDirectMessageChannels lists direct message channels
+func (a *API) ListDirectMessageChannels() ([]*Channel, error) {
+	data, err := newRequest(IMList).
+		addParam("token", a.botToken).
+		Get()
+	if err != nil {
+		return nil, errors.Wrap(err, "api request")
+	}
+	resp := &listDirectMessageChannelsReponse{}
+	if err := json.Unmarshal(data, resp); err != nil {
+		return nil, errors.Wrap(err, "unmarshal")
+	}
+	if !resp.OK {
+		return nil, errors.New(resp.Error)
+	}
+	return resp.Channels, nil
+}
+
+
+
+type deleteMessageResponse struct {
+	OK    bool   `json:"ok"`
+	Error string `json:"error"`
+}
+
+// DeleteBotMessage deletes a bot message
+func (a *API) DeleteBotMessage(channelID string, messageTS string) error {
+	data, err := newRequest(ChatDelete).
+		addParam("channel", channelID).
+		addParam("ts", messageTS).
+		addParam("token", a.botToken).
+		Get()
+	if err != nil {
+		return errors.Wrap(err, "api request")
+	}
+	resp := &deleteMessageResponse{}
+	if err := json.Unmarshal(data, resp); err != nil {
+		return errors.Wrap(err, "unmarshal")
+	}
+	if !resp.OK {
+		return errors.New(resp.Error)
+	}
+	return nil
 }
